Rename module 'api' to 'api_c'
