+++ libvcx/src/api_c/connection.rs
@@ -1211,7 +1211,7 @@ mod tests {
     use crate::abi_utils::return_types_u32;
     use crate::abi_utils::timeout::TimeoutUtils;
     use crate::agency_client::mocking::AgencyMockDecrypted;
-    use crate::api::VcxStateType;
+    use crate::api_c::VcxStateType;
     use crate::connection::tests::{build_test_connection_inviter_invited, build_test_connection_inviter_null, build_test_connection_inviter_requested};
     use crate::utils::constants::{DELETE_CONNECTION_DECRYPTED_RESPONSE, GET_MESSAGES_DECRYPTED_RESPONSE};
     use crate::utils::devsetup::*;
