+++ libvcx/src/api_c/issuer_credential.rs
@@ -851,7 +851,7 @@ pub mod tests {
 
     use crate::abi_utils::return_types_u32;
     use crate::abi_utils::timeout::TimeoutUtils;
-    use crate::api::VcxStateType;
+    use crate::api_c::VcxStateType;
     use crate::settings;
     use crate::utils::constants::*;
     use crate::utils::devsetup::*;
