+++ libvcx/src/api_c/utils.rs
