+++ libvcx/src/api_c/wallet.rs
