+++ libvcx/src/proof.rs
@@ -123,7 +123,7 @@ pub mod tests {
     use agency_client::mocking::HttpClientMockResponse;
 
     use crate::proof;
-    use crate::api::VcxStateType;
+    use crate::api_c::VcxStateType;
     use crate::aries::handlers::proof_presentation::verifier::verifier::Verifier;
     use crate::aries::messages::proof_presentation::presentation::tests::_comment;
     use crate::connection::tests::build_test_connection_inviter_requested;
